* (1) Changed semantics of Cell::body(), Cell::bodies(), Cell::body_attr(), Cell::body_attrs() (2) fixed a few const-quaifier issues (3) fixed direct computation check in ExaFMM's serial_tapas binary
Semantics of Cell::body(), etc. is "all bodies" and it works for the single node HOT implementation.
However, it does not work for distributed version because returning all bodies is impossible.
Thus, changed the seamntics of Cell::body() to "returns local bodies" and introduced nbodies() function,
which returns the number of local bodies in the current process.
This design decision is temporary and remain to be improved in future.

The change fixes check calculation in distributed version (of 1 process so far).
It was broken in parallel_tapas_mt binary because it copied the computed results
to user's original array of bodies. In parallel version, bodies are distributed and sorted
and no longer correspond to user's original body array.